pyproject.toml
[project]
name = "zwy-stock-service"
version = "0.1.0"
description = "Project scaffold generated by init_fastapi_project.sh"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "apscheduler>=3.10.4",
  "httpx>=0.27.2",
  "bcrypt>=4.2.0",
  "celery>=5.4.0",
  "fastapi>=0.115.0",
  "loguru>=0.7.2",
  "pydantic-settings>=2.5.2",
  "python-dateutil>=2.9.0.post0",
  "pytz>=2024.1",
  "redis>=5.0.8",
  "shortuuid>=1.0.13",
  "sqlalchemy>=2.0.36",
  "asyncmy>=0.2.9",
  "asyncpg>=0.30.0",
  "uvicorn[standard]>=0.30.6",
  "email-validator>=2.3.0",
]

[dependency-groups]
dev = [
  "alembic>=1.14.1",
  "pytest>=8.3.2",
]

[tool.uv]
package = false

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"